Back-off an unneeded cast added by mistake in the 1.7.4.1 MFC
Former-commit-id: b4c3a20292e1a8e49f5fffd211dd1e7c2bb0ed76

+++ secure/usr.bin/bdes/bdes.c
@@ -195,7 +195,7 @@ parameter_list|,
 name|n
 parameter_list|)
 define|\
-value|if (fwrite(buf, sizeof(char), n, stdout) != (size_t)n)	\ 			warnx("fwrite error at %d", n);
+value|if (fwrite(buf, sizeof(char), n, stdout) != n)	\ 			warnx("fwrite error at %d", n);
 end_define
 
 begin_comment
